% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl.var.global.R
\docType{data}
\name{expl.var.global}
\alias{expl.var.global}
\title{Data expl.var.global}
\format{
\subsection{\code{expl.var.global}}{

A \code{SpatRaster} object with the following layers:
\describe{
\item{bio1}{Annual Mean Temperature}
\item{bio2}{Mean Diurnal Range (Mean of monthly (max temp - min temp))}
\item{bio4}{Temperature Seasonality (standard deviation ×100)}
\item{bio12}{Annual Precipitation}
}
}
}
\source{
\url{https://nsojournals.onlinelibrary.wiley.com/doi/full/10.1111/ecog.07328}
}
\usage{
expl.var.global
}
\description{
A dataset containing environmental covariates (usually climatic) at global level.
}
\keyword{datasets}
