% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fagus.sylvatica.xy.global.R
\docType{data}
\name{Fagus.sylvatica.xy.global}
\alias{Fagus.sylvatica.xy.global}
\title{Data Fagus.sylvatica.xy.global}
\format{
\subsection{\code{Fagus.sylvatica.xy.global}}{

A \code{data.frame} with 7,205 rows and 2 columns:
\describe{
\item{x}{Longitude of the occurrence point}
\item{y}{Latitude of the occurrence point}
}
}
}
\source{
\url{https://nsojournals.onlinelibrary.wiley.com/doi/full/10.1111/ecog.07328}
}
\usage{
Fagus.sylvatica.xy.global
}
\description{
A dataset containing the coordinates (x, y) of the occurrence points of the species (\emph{Fagus sylvatica} at global level.
}
\keyword{datasets}
