% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fagus.sylvatica.xy.regional.R
\docType{data}
\name{Fagus.sylvatica.xy.regional}
\alias{Fagus.sylvatica.xy.regional}
\title{Regional Occurrences of Fagus sylvatica}
\format{
A \code{data.frame} with 2,011 rows and 2 columns:
\describe{
\item{\code{Longitude}}{Longitude of occurrence points.}
\item{\code{Latitude}}{Latitude of occurrence points.}
}
}
\source{
\doi{10.1111/ecog.07328}
}
\usage{
data(Fagus.sylvatica.xy.regional)
}
\description{
A dataset containing the geographic coordinates of occurrence points for \emph{Fagus sylvatica} at a regional scale.
}
\keyword{datasets}
