% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalModel.R
\name{NSDM.Global}
\alias{NSDM.Global}
\title{Perform nested spatial hierarchical species distribution modeling (NSDM) analysis at global scale.}
\usage{
NSDM.Global(
  nsdm_selvars,
  algorithms = c("GLM", "GAM", "RF"),
  CV.nb.rep = 10,
  CV.perc = 0.8,
  CustomModelOptions = NULL,
  metric.select.thresh = 0.8,
  save.output = TRUE,
  rm.biomod.folder = TRUE
)
}
\arguments{
\item{nsdm_selvars}{An object of class \code{nsdm.vinput} containing selected covariables for NSDM generated using the \code{\link{NSDM.SelectCovariates}} function.}

\item{algorithms}{(\emph{optional, default} \code{'c("GLM", "GAM", "RF")'}) \cr
Algorithms to use for modeling. Options are \code{'GLM'}, \code{'GAM'}, \code{'GBM'}, \code{'MAXNET'}, \code{'MARS'}, and/or \code{'RF'}.}

\item{CV.nb.rep}{(\emph{optional, default} \code{10}) \cr
An \code{integer} corresponding to the number of sets (repetitions) of cross-validation points that will be drawn.}

\item{CV.perc}{(\emph{optional, default} \code{0.8}) \cr
A \code{numeric} between \code{0} and \code{1} defining the percentage of data that will be kept for calibration.}

\item{CustomModelOptions}{(\emph{optional, default} \code{NULL}) \cr
A \code{\link{BIOMOD.models.options}} object returned by the \code{\link{bm_ModelingOptions}}}

\item{metric.select.thresh}{(\emph{optional, default} \code{0.8}) \cr
A \code{numeric} between \code{0} and \code{1} corresponding the minimum scores of AUC below which single models will be excluded from the ensemble model building.}

\item{save.output}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} value defining whether the outputs should be saved at local.}

\item{rm.biomod.folder}{(\emph{optional, default} \code{TRUE}) \cr
A \code{logical} value indicating whether the intermediate BIOMOD2 folders should be removed after processing.}
}
\value{
An object of class \code{nsdm.predict.g} containing model information, predictions and evaluation statistics:
\itemize{
\item \verb{$SpeciesName} Name of the species.
\item \verb{$SpeciesData.XY.Global} Species presence data at the global level at \code{data.frame} format after applying spatial thinning.
\item \verb{$SpeciesData.XY.Regional} Species presence data at the regional level at \code{data.frame} format after applying spatial thinning.
\item \verb{$Background.XY.Global} Background data at the global level at \code{data.frame} format.
\item \verb{$Background.XY.Regional} Species presence data at the regional level at \code{data.frame} format.
\item \verb{$Scenarios} A \code{list} containing future scenarios in \code{\link[terra:rast]{PackedSpatRaster}} format.
\item \verb{$Selected.Variables.Global} A \code{character} vector specifying the names of the selected covariables at the global scale.
\item \verb{$IndVar.Global.Selected} Selected independent variables at the global level in \code{\link[terra:rast]{PackedSpatRaster}} format.
\item \verb{$Selected.Variables.Regional} A \code{character} vector specifying the names of the selected covariables at the regional scale.
\item \verb{$IndVar.Regional.Selected} Selected independent variables at the regional level in \code{\link[terra:rast]{PackedSpatRaster}} format.
\item \verb{$IndVar.Global.Selected.reg} Selected variables at the global level for regional projections in \code{\link[terra:rast]{PackedSpatRaster}} format.
\item \verb{$args} A \code{list} containing the arguments used during modelling, including: \code{algorithms}, \code{CV.nb.rep}, \code{CV.perc} and \code{metric.select.thresh}.
\item \verb{$nbestreplicates} A \code{data.frame} containing  the number of replicates meeting or exceeding the specified \code{metric.select.thresh} for each algorithm used in the modeling.
\item \verb{$current.projections} A \code{list} containing: \code{Pred}, a \code{\link[terra:rast]{PackedSpatRaster}} representing the current projection.....; \code{Pred.bin.ROC}, a \code{\link[terra:rast]{PackedSpatRaster}} representing projections ..........; and \code{Pred.bin.TSS}, a \code{\link[terra:rast]{PackedSpatRaster}} representing......
\item \verb{$myEMeval.replicates} Evaluation statistics for each replicate model
\item \verb{$myEMeval.Ensemble} Evaluation statistics for the ensemble model.
\item \verb{$myModelsVarImport} Variable importance measures for individual models.
\item \verb{$new.projections} A \code{list} containing: \code{Pred.Scenario}, the projections onto new scenarios in a \code{\link[terra:rast]{PackedSpatRaster}} format; \code{Pred.bin.ROC.Scenario}, the binary projections onto new scenarios in a \code{\link[terra:rast]{PackedSpatRaster}} format, derived from ROC scores; and \code{Pred.bin.TSS.Scenario}, the binary projections onto new scenarios in a \code{\link[terra:rast]{PackedSpatRaster}} format, derived from ROC scores.
\item \code{Summary} Summary information about the modeling process.
}
}
\description{
This function conducts \bold{NSDM} analysis at the global scale. It trains individual models using specified algorithms, evaluates model performance, and projects models to current and new environmental conditions.
}
\details{
This function conducts \bold{NSDM} modeling at the global scale. It uses the biomod2 package.
If \code{save.output=TRUE}, modelling results are stored out of R in the \emph{Results/} folder created in the current working directory:
\itemize{
\item the \emph{Results/Global/Projections/} folder, containing the continious and binary current and new projections. Current projections are named with the species name followed by \file{.Current.tif}, \file{.bin.ROC.tif} and \file{.bin.TSS.tif}. New projections are named with the species name followed by the scenario name, and \file{.bin.ROC.tif}, \file{.bin.TSS.tif} when binary.
\item the \emph{Results/Global/Values/} folder, containing replicates statistics, the consensus model statistics, the covariable importance, and the \code{nbestreplicates}, named with the species name and \file{.__replica.csv}, \file{._ensemble.csv}, \file{._indvar.csv} and \file{._nbestreplicates.csv} respectively.
}
}
\examples{
# Perform NSDM modeling at global scale  #@@JMB Ver cómo hacen otros cuando una función depende de objetos anteriores
myGlobalModel <- NSDM.Global(mySelectedCovs)

}
\seealso{
\code{\link{NSDM.InputData}}, \code{\link{NSDM.FormattingData}}, \code{\link{NSDM.SelectCovariates}}
}
