% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{NSDM.InputData}
\alias{NSDM.InputData}
\title{Prepare input data for nested spatial hierarchical species distribution modeling (NSDM) analysis.}
\usage{
NSDM.InputData(
  SpeciesName,
  spp.data.global,
  spp.data.regional,
  expl.var.global,
  expl.var.regional,
  new.env = NULL,
  new.env.names = NULL,
  Background.Global = NULL,
  Background.Regional = NULL
)
}
\arguments{
\item{SpeciesName}{A \code{character} specifying the species name for which data is being prepared.}

\item{spp.data.global}{A \code{data.frame} with two columns, 'x' and 'y', representing the species presence data at global level.}

\item{spp.data.regional}{A \code{data.frame} with two columns, 'x' and 'y', representing the species presence data at regional level.}

\item{expl.var.global}{An object of \code{\link[terra:rast]{SpatRaster}} class representing the environmental covariates (usually climatic) at global level.}

\item{expl.var.regional}{An object of \code{\link[terra:rast]{SpatRaster}} class representing the environmental covariates at regional level.}

\item{new.env}{(\emph{optional, default} \code{NULL}) \cr
An object of \code{\link[terra:rast]{SpatRaster}} class or a \code{list} of \code{\link[terra:rast]{SpatRaster}} objects representing new environmental scenarios for projecting the models to different spatial or temporal scales.}

\item{new.env.names}{(\emph{optional, default} \code{NULL}) \cr
A character \code{vector} specifying names for the new environmental scenarios.}

\item{Background.Global}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing background points at global level.}

\item{Background.Regional}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing background points at regional level.}
}
\value{
An object of class "nsdm.input" containing organized input data for NSDM.
}
\description{
This function gathers together all input data (\emph{species presences xy at global and regional level, enviromental covariates at global and regional level, new environments, and if available, absence/pseudo-absences/background data}) needed to run \bold{NSDM}.
}
\details{
\itemize{
\item The \code{SpeciesName} parameter specifies the name of the species for which data is being prepared.
\item The \code{spp.data.global} and \code{spp.data.regional} parameters should be \code{data.frame} with columns 'x' and 'y' representing species presence data.
\item The \code{expl.var.global} and \code{expl.var.regional} parameters should be \code{\link[terra:rast]{SpatRaster}} objects representing enviromental covariates.  Each band correpond to a different covariate. At the global-scale, only one \code{\link[terra:rast]{SpatRaster}} file must be provided corresponding to the covariates (usually climatic) used to train the global-scale model. At the regional-scale, there must be at least one geotiff file must corresponding to the covariates used to train the regional-scale model. The regional-scale \code{\link[terra:rast]{SpatRaster}} file must include all the covariates included in the global-scale file, and it can additionally include covariates only available at this level.
\item The \code{new.env} parameter can be either a \code{\link[terra:rast]{SpatRaster}} object or a \code{list} of \code{\link[terra:rast]{SpatRaster}} objects representing new environmental scenarios.
\item The \code{Background.Global} and \code{Background.Regional} parameters are optional and represent background points at the global and regional levels, respectively.
}
}
\examples{
library(terra)

# Load species occurrences
data(Fagus.sylvatica.xy.global, package = "sabinaNSDM")
data(Fagus.sylvatica.xy.regional, package = "sabinaNSDM")

# Load explanatory variables
data(expl.var.global, package = "sabinaNSDM")
data(expl.var.regional, package = "sabinaNSDM")
expl.var.global<-terra::unwrap(expl.var.global)
expl.var.regional<-terra::unwrap(expl.var.regional)

# Load new escenarios
data(new.env, package = "sabinaNSDM")
new.env<-terra::unwrap(new.env)

myInputData<-NSDM.InputData(
	SpeciesName = "Fagus.sylvatica",
	spp.data.global = Fagus.sylvatica.xy.global,
	spp.data.regional = Fagus.sylvatica.xy.regional,
	expl.var.global = expl.var.global,
	expl.var.regional = expl.var.regional,
	new.env = new_env,
	new.env.names = c("Scenario1"),
	Background.Global = NULL,
	Background.Regional = NULL
)


}
