% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quercus.petraea.xy.global.R
\docType{data}
\name{Quercus.petraea.xy.global}
\alias{Quercus.petraea.xy.global}
\title{Data Quercus.petraea.xy.global}
\format{
\subsection{\code{Quercus.petraea.xy.global}}{

A \code{data.frame} with 4,364 rows and 2 columns:
\describe{
\item{x}{Longitude of the occurrence point}
\item{y}{Latitude of the occurrence point}
}
}
}
\source{
\url{https://nsojournals.onlinelibrary.wiley.com/doi/full/10.1111/ecog.07328}
}
\usage{
Quercus.petraea.xy.global
}
\description{
A dataset containing the coordinates (x, y) of the occurrence points of the species (\emph{Quercus petraea} at global level.
}
\keyword{datasets}
