% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quercus.petraea.xy.regional.R
\docType{data}
\name{Quercus.petraea.xy.regional}
\alias{Quercus.petraea.xy.regional}
\title{Data Quercus.petraea.xy.regional}
\format{
\subsection{\code{Quercus.petraea.xy.regional}}{

A \code{data.frame} with 775 rows and 2 columns:
\describe{
\item{x}{Longitude of the occurrence point}
\item{y}{Latitude of the occurrence point}
}
}
}
\source{
\url{https://nsojournals.onlinelibrary.wiley.com/doi/full/10.1111/ecog.07328}
}
\usage{
Quercus.petraea.xy.regional
}
\description{
A dataset containing the coordinates (x, y) of the occurrence points of the species (\emph{Quercus petraea} at regional level.
}
\keyword{datasets}
