% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{summary.nsdm.input}
\alias{summary.nsdm.input}
\title{Summary for nsdm.input Objects}
\usage{
\method{summary}{nsdm.input}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nsdm.input}.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary in a data frame format.
}
\description{
Provides a summary for objects of class \code{nsdm.input}.
}
\seealso{
\code{\link{NSDM.InputData}} \code{\link{NSDM.FormattingData}} \code{\link{NSDM.SelectCovariates}}
}
