% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{summary.nsdm.predict}
\alias{summary.nsdm.predict}
\title{Summary for nsdm.predict Objects}
\usage{
\method{summary}{nsdm.predict}(object, ...)
}
\arguments{
\item{object}{An object of class \code{nsdm.predict}.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary in a data frame format.
}
\description{
Provides a summary for objects of class \code{nsdm.predict}.
}
\seealso{
\code{\link{NSDM.Global}} \code{\link{NSDM.Regional}} \code{\link{NSDM.Covariate}} \code{\link{NSDM.Multiply}}
}
