% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expl.var.regional.R
\docType{data}
\name{expl.var.regional}
\alias{expl.var.regional}
\title{Data expl.var.regional}
\format{
\subsection{\code{expl.var.regional}}{

A \code{SpatRaster} object with the following layers:
\describe{
\item{bio1}{Annual Mean Temperature}
\item{bio2}{Mean Diurnal Range (Mean of monthly (max temp - min temp))}
\item{bio3}{Isothermality (BIO2/BIO7) (×100)}
\item{bio4}{Temperature Seasonality (standard deviation ×100)}
\item{bio12}{Annual Precipitation}
\item{sand}{Sand content}
\item{radiation}{Annual solar radiation}
}
}
}
\source{
\url{https://nsojournals.onlinelibrary.wiley.com/doi/full/10.1111/ecog.07328}
}
\usage{
expl.var.regional
}
\description{
A dataset containing multiple environmental covariates at a regional level.
}
\keyword{datasets}
