% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputData.R
\name{NSDM.InputData}
\alias{NSDM.InputData}
\title{Prepare input data for spatially-nested hierarchical species distribution modeling (NSDM) analysis.}
\usage{
NSDM.InputData(
  SpeciesName,
  spp.data.global,
  spp.data.regional,
  expl.var.global,
  expl.var.regional,
  new.env = NULL,
  new.env.names = NULL,
  Background.Global = NULL,
  Background.Regional = NULL,
  Absences.Global = NULL,
  Absences.Regional = NULL
)
}
\arguments{
\item{SpeciesName}{A \code{character} specifying the species name for which data is being prepared.}

\item{spp.data.global}{A \code{data.frame} with two columns, 'x' and 'y', representing the species presence data at global level.}

\item{spp.data.regional}{A \code{data.frame} with two columns, 'x' and 'y', representing the species presence data at regional level.}

\item{expl.var.global}{An object of \code{\link[terra:rast]{SpatRaster}} class representing the environmental covariates (usually climatic) at global level.}

\item{expl.var.regional}{An object of \code{\link[terra:rast]{SpatRaster}} class representing the environmental covariates at regional level.}

\item{new.env}{(\emph{optional, default} \code{NULL}) \cr
An object of \code{\link[terra:rast]{SpatRaster}} class or a \code{list} of \code{\link[terra:rast]{SpatRaster}} objects representing the environmental covariates in new scenarios for projecting the models to different spatial or temporal scales.}

\item{new.env.names}{(\emph{optional, default} \code{NULL}) \cr
A character \code{vector} specifying names for the new environmental scenarios.}

\item{Background.Global}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing background points at global level.}

\item{Background.Regional}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing background points at regional level.}

\item{Absences.Global}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing absence points at global level.}

\item{Absences.Regional}{(\emph{optional, default} \code{NULL}) \cr
An optional \code{data.frame} with two columns, 'x' and 'y', representing absence points at regional level.}
}
\value{
An object of class \code{nsdm.input} containing organized input data for \bold{NSDM}.
}
\description{
This function gathers together all input data (\emph{species occurrences xy at global and regional level, enviromental covariates at global and regional level, new environments, and if available, absence/pseudo-absences/background data}) needed to run \bold{NSDM}.
}
\details{
\itemize{
\item The \code{expl.var.global} and \code{expl.var.regional} parameters should be \code{\link[terra:rast]{SpatRaster}} objects representing environmental covariates used to train the global-scale and regional-scale models, respectively. Each band corresponds to a different covariate. The regional-scale object must include all the covariates included in the global-scale object (usually climatic), and it can additionally include other covariates only available at this level.
\item The \code{new.env} parameter can be either a \code{\link[terra:rast]{SpatRaster}} object or a \code{list} of \code{\link[terra:rast]{SpatRaster}} object representing the environmental covariates in new spatial or temporal scenarios. All new scenarios must have the same covariates (same band names) than \code{expl.var.regional}.
}
}
\examples{
library(sabinaNSDM)

# Load species occurrences
data(Fagus.sylvatica.xy.global, package = "sabinaNSDM")
data(Fagus.sylvatica.xy.regional, package = "sabinaNSDM")

# Load covariates
data(expl.var.global, package = "sabinaNSDM")
data(expl.var.regional, package = "sabinaNSDM")
expl.var.global<-terra::unwrap(expl.var.global)
expl.var.regional<-terra::unwrap(expl.var.regional)

# Load new scenarios
data(new.env, package = "sabinaNSDM")
new.env<-terra::unwrap(new.env)

# Prepare input data
myInputData<-NSDM.InputData(SpeciesName = "Fagus.sylvatica",
			spp.data.global = Fagus.sylvatica.xy.global,
			spp.data.regional = Fagus.sylvatica.xy.regional,
			expl.var.global = expl.var.global,
			expl.var.regional = expl.var.regional,
			new.env = new.env,
			new.env.names = c("Scenario1"),
			Background.Global = NULL,
			Background.Regional = NULL,
			Absences.Global = NULL,
			Absences.Regional = NULL)

summary(myInputData)


## ------------------------------------------------------------------
## Example: multiple new scenarios
## Replicate the new scenario to simulate a second scenario
# scenario1 <- new.env
# scenario2 <- new.env  # Replicate to create a second scenario

## Combine scenarios into a list
# new_scenarios <- list(scenario1, scenario2)
# new_scenario_names <- c("Scenario1", "Scenario2")

## Prepare input data with multiple scenarios
# myInputData_multi <- NSDM.InputData(SpeciesName = "Fagus.sylvatica",
#                                     spp.data.global = Fagus.sylvatica.xy.global,
#                                     spp.data.regional = Fagus.sylvatica.xy.regional,
#                                     expl.var.global = expl.var.global,
#                                     expl.var.regional = expl.var.regional,
#                                     new.env = new_scenarios,
#                                     new.env.names = new_scenario_names,
#                                     Background.Global = NULL,
#                                     Background.Regional = NULL,
#                                     Absences.Global = NULL,
#                                     Absences.Regional = NULL)


## ------------------------------------------------------------------
## Example: running a single model (non-nested workflow)
## Provide only regional data (leave global = NULL)

# myInputData_ref<- NSDM.InputData(SpeciesName = "Fagus.sylvatica",
#                                  spp.data.global = NULL,
#                                  spp.data.regional = Fagus.sylvatica.xy.regional,
#                                  expl.var.global = NULL,
#                                  expl.var.regional = expl.var.regional,
#                                  Background.Global = NULL,
#                                  Background.Regional = NULL,
#                                  Absences.Global = NULL,
#                                  Absences.Regional = NULL)

}
