% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quercus.petraea.xy.global.R
\docType{data}
\name{Quercus.petraea.xy.global}
\alias{Quercus.petraea.xy.global}
\title{Global Occurrences of Quercus petraea}
\format{
A \code{data.frame} with 4,364 rows and 2 columns:
\describe{
\item{\code{Longitude}}{Longitude of species occurrence points.}
\item{\code{Latitude}}{Latitude of species occurrence points.}
}
}
\source{
\doi{10.1111/ecog.07328}
}
\usage{
data(Quercus.petraea.xy.global)
}
\description{
A dataset containing the coordinates (x, y) of the occurrence points of the species (\emph{Quercus petraea} at global level.
}
\keyword{datasets}
